requirements.txt
pandas
bs4
flask
requests
PyYAML
pytest